refactor: Change BalmCustomConfigControlRegistrar to be backed by an interface, move ConfigControlRegistry to internal package
