+++ common/src/main/java/net/blay09/mods/balm/platform/compatibility/config/internal/ClothConfigSupport.java
@@ -5,7 +5,7 @@ import me.shedaniel.clothconfig2.api.ConfigBuilder;
 import me.shedaniel.clothconfig2.api.ConfigCategory;
 import net.blay09.mods.balm.Balm;
 import net.blay09.mods.balm.client.platform.config.BalmConfigScreenFactory;
-import net.blay09.mods.balm.client.platform.config.ConfigControlRegistry;
+import net.blay09.mods.balm.client.platform.config.internal.ConfigControlRegistry;
 import net.blay09.mods.balm.platform.config.MutableLoadedConfig;
 import net.blay09.mods.balm.platform.config.internal.BalmConfigScreenProviders;
 import net.blay09.mods.balm.platform.config.schema.*;
