+++ common/src/main/java/net/blay09/mods/balm/platform/compatibility/config/internal/ConfiguredConfigProvider.java
@@ -4,7 +4,7 @@ import com.mrcrayfish.configured.api.*;
 import com.mrcrayfish.configured.api.util.ConfigScreenHelper;
 import net.blay09.mods.balm.Balm;
 import net.blay09.mods.balm.client.platform.config.BalmConfigScreenFactory;
-import net.blay09.mods.balm.client.platform.config.ConfigControlRegistry;
+import net.blay09.mods.balm.client.platform.config.internal.ConfigControlRegistry;
 import net.blay09.mods.balm.platform.config.MutableLoadedConfig;
 import net.blay09.mods.balm.platform.config.schema.BalmConfigSchema;
 import net.blay09.mods.balm.platform.config.schema.ConfigControlContext;
